refactor: useCategory 커스텀훅
